requirements.txt
# MultiAgent Operations — Python dependencies
# Generated from .venv — Python 3.14
# Install: python3 -m venv .venv && source .venv/bin/activate && pip install -r requirements.txt

# Web framework + production server
Flask==3.1.3
gunicorn==21.2.0

# Data
pandas==3.0.1
numpy==2.4.2

# Discord
discord.py==2.7.1

# Environment / config
python-dotenv==1.2.2

# Google / Gmail API
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-oauthlib==1.3.0
google-auth-httplib2==0.3.0

# Task scheduling
schedule==1.2.2

# Google Ads API (Best Care integration)
google-ads==25.1.0

# HTTP requests (competitor intel, 8x8 fallback)
requests==2.32.3
